+++ test/unit/dropin.js
@@ -2285,4 +2285,43 @@ describe('Dropin', () => {
       });
     });
   });
+
+  describe('3ds events', () => {
+    test.each([
+      'customer-canceled',
+      'authentication-modal-render',
+      'authentication-modal-close'
+    ])('emits 3ds:%s event', (eventName, done) => {
+      const instance = new Dropin(testContext.dropinOptions);
+      const payload = {};
+
+      instance.on(`3ds:${eventName}`, (emittedPayload) => {
+        expect(emittedPayload).toBe(payload);
+        done();
+      });
+
+      instance._initialize(() => {
+        instance._model._emit(`3ds:${eventName}`, payload);
+      });
+    });
+  });
+
+  describe('payment option selected event', () => {
+    test(
+      'emits paymentOptionSelected when the model emits paymentOptionSelected',
+      done => {
+        const instance = new Dropin(testContext.dropinOptions);
+
+        instance.on('paymentOptionSelected', event => {
+          expect(event.paymentOption).toBe('Foo');
+
+          done();
+        });
+
+        instance._initialize(() => {
+          instance._model._emit('paymentOptionSelected', { paymentOption: 'Foo' });
+        });
+      }
+    );
+  });
 });