+++ CHANGELOG.md
@@ -8,6 +8,10 @@ unreleased
 - Add `changeActiveView` event to track when the Drop-in updates the active view presented to the customer
 - Google Pay
   - Fix issue where passing custom button option would caused a developer error in the console (#701)
+- 3D Secure
+  - Add `3ds:customer-canceled` event (#690)
+  - Add `3ds:authentication-modal-render` event
+  - Add `3ds:authentication-modal-close` event
 
 1.26.1
 ------